changed CLI behavior, format of csv outputs
