+++ README.md
@@ -4,7 +4,9 @@ A command line tool to help review and reconcile Naxos records
 
 ## Commands
 
-### `prep`
+All commands can be run 
+
+### `naxos prep`
 Process `.xml` files from Naxos and/or `.csv` file from Sierra export
 
 #### Options:
@@ -28,7 +30,7 @@ Naxos:
     - Control Number (from 001 field)
     - CID (from URL)
 
-### `compare`
+### `naxos compare`
 Compare prepped Naxos and Sierra files
 
 #### Options
@@ -43,7 +45,7 @@ Compare prepped Naxos and Sierra files
 5) Create dataframe for all resources only present in Sierra data. Export results to "records_to_delete.csv"
 6) Create dataframe for all resources only present in Naxos data. Export results to "records_to_import.csv"
 
-### `reconcile`
+### `naxos reconcile`
 Prep files from Sierra and Naxos and then compare them
 
 #### Options
@@ -54,13 +56,13 @@ Prep files from Sierra and Naxos and then compare them
 1) Prepares files using process outlined above in `prep`
 2) Compares files using process outlined above in `compare`
 
-### `check-urls`
+### `naxos check-urls`
 Check URLs for each row in a spreadsheet; data should be prepped first
 
 #### Options
 `-f` `--file`: file to check
 
 
-### `sample`
+### `naxos sample`
 Create a sample of data from a spreadsheet
